+++ tests/stubs/CustomAsset.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Thinktomorrow\AssetLibrary\Tests\stubs;
+
+use Thinktomorrow\AssetLibrary\AbstractAsset;
+
+class CustomAsset extends AbstractAsset
+{
+
+}